pyproject.toml
[project]
name = "copier-turbo-django"
version = "2025.12.05"
description = "A Copier template for creating production-ready modular monolithic Django projects with Turborepo."
readme = "README.md"
keywords = [
  "copier",
  "django",
  "project template",
  "scaffolding",
  "skeleton",
]
license = { text = "BSD" }
authors = [
  { name = "Mike Ascah", email = "mike@ascah.dev" },
]
requires-python = "==3.13.*"
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Framework :: Django :: 5.2",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Natural Language :: English",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Software Development",
]
dependencies = [
  "binaryornot==0.4.4",
  "copier>=9.0.0",
  "django-upgrade==1.29.1",
  "djlint==1.36.4",
  "gitpython==3.1.45",
  "jinja2==3.1.6",
  "jinja2-time>=0.2.0",
  "pre-commit==4.5.0",
  "pygithub==2.8.1",
  "pytest==9.0.2",
  "pytest-instafail==0.5",
  "pytest-xdist==3.8.0",
  "pyyaml==6.0.3",
  "requests==2.32.5",
  "ruff==0.14.8",
  "sh==2.2.2; sys_platform!='win23'",
  "tox==4.32.0",
  "tox-uv>=1.17",
]
urls = { Repository = "https://github.com/mascah/cookiecutter-turbo-django" }

[dependency-groups]
docs = [
  "myst-parser>=4",
  "pygments>=2.19",
  "sphinx>=8.0.2,<9",
  "sphinx-autobuild>=2024.10.3",
  "sphinx-rtd-theme>=3",
]

[tool.ruff]
target-version = "py39"
line-length = 119
# Exclude the template content as most files aren't parseable
extend-exclude = [
  "template/*",
  "docs/*",
]

lint.select = [
  "A",
  # "ANN", # flake8-annotations: we should support this in the future but many errors atm
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FBT",
  "FLY",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "N",
  "PD",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  # "ARG", # Unused function argument
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  # "FURB",
  # "LOG",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  "EM101",
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "SIM102", # sometimes it's better to nest
  "TRY003", # Avoid specifying long messages outside the exception class
]
lint.isort.force-single-line = true

[tool.pyproject-fmt]
keep_full_version = true

# ==== pytest ====

[tool.pytest.ini_options]
addopts = "-v --tb=short"
norecursedirs = [
  ".tox",
  ".git",
  "*/migrations/*",
  "*/static/*",
  "docs",
  "venv",
  "ref",
  "template",
]

# ==== djLint ====

[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002,T028"
ignore_blocks = "raw"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "jinja"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2
